requirements.txt
Arpeggio==2.0.2
attrs==24.2.0
azure-core==1.32.0
azure-functions==1.21.3
azure-storage-blob==12.18.0
azure-storage-queue==12.7.0
boto3==1.35.33
botocore==1.35.33
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==44.0.0
debugpy==1.8.6
Deprecated==1.2.14
dill==0.3.9
distlib==0.3.8
exceptiongroup==1.2.2
filelock==3.15.4
googleapis-common-protos==1.63.2
grpcio==1.60.2
idna==3.7
importlib_metadata==8.0.0
iniconfig==2.0.0
isodate==0.7.2
jmespath==1.0.1
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
packaging==24.1
parver==0.5
platformdirs==4.2.2
pluggy==1.5.0
protobuf==4.25.4
psutil==6.0.0
pulumi==3.135.1
pulumi_aws==6.54.2
pycparser==2.22
pytest==7.4.0
pytest-opentelemetry==0.1.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.31.0
s3transfer==0.10.2
semver==2.13.0
services==0.1.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
wrapt==1.16.0
zipp==3.19.2
xmltodict==0.14.2
